Patched **kwargs inconsistencies
This should solve issue #332.
Also added rescaling mecanism for line to avoid line clipping out of header in waveform figures.

This triggers per-line (only the line that would overflow is rescaled). I tested it on the mtuq examples and it looks okay to me.

Also tested with plot_data_greens3 and everything looks fine as well. I turned it on by default, but we could change that if needed.